refactor: move most of cli processing to a cli crate
