+++ crates/monotax/src/main.rs
@@ -1,12 +1,14 @@
 use std::process::exit;
 
 use clap::Parser;
+use cli::app::run_cli_command;
+use cli::Cli;
 use env_logger::{Builder, Env};
 use log::{debug, error};
-use monotax_core::cli::app::run_cli_command;
-use monotax_core::cli::Cli;
 use monotax_core::config::load_config;
 use monotax_core::infra::sqlx::connection::connection_pool;
+
+mod cli;
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
