+++ crates/monotax_core/src/lib.rs
@@ -5,7 +5,6 @@
 //! The central concept of Monotax is an income. An income is a record of money received by a person.
 
 pub mod app;
-pub mod cli;
 pub mod config;
 pub mod domain;
 pub mod filter;
